# Conflicts:
#	cppmodule/deps/pybind11

+++ cppcore/tests/test_modifiers.cpp
@@ -205,9 +205,9 @@ TEST_CASE("SiteGenerator") {
         auto const energy = MatrixXcd::Constant(1, 1, 2.0);
         model.add(SiteGenerator("C", energy, [](System const&) {
             auto const size = 5;
-            auto x = ArrayXf::Constant(size, 1);
-            auto y = ArrayXf::LinSpaced(size, 1, 5);
-            auto z = ArrayXf::Constant(size, 0);
+            auto x = ArrayXd::Constant(size, 1);
+            auto y = ArrayXd::LinSpaced(size, 1, 5);
+            auto z = ArrayXd::Constant(size, 0);
             return CartesianArray(x, y, z);
         }));
